+++ ptolemy/actor/lib/jjs/NashornAccessorHostApplication.java
@@ -410,7 +410,7 @@ public class NashornAccessorHostApplication {
          *  @return handle A handle that can be used to cancel the timeout.
          */
         @Override
-        public synchronized Timer setInterval(Runnable function, int periodMS) {
+        public synchronized Timer setInterval(Runnable function, long periodMS) {
             // System.out.println(_name + ": requesting interval: " + periodMS + "ms for " + function + " in thread " + Thread.currentThread());
             TimerTask task = new TimerTask() {
                 @Override
@@ -433,7 +433,7 @@ public class NashornAccessorHostApplication {
          *  @return handle A handle that can be used to cancel the timeout.
          */
         @Override
-        public synchronized Timer setTimeout(Runnable function, int timeMS) {
+        public synchronized Timer setTimeout(Runnable function, long timeMS) {
             // System.out.println(_name + ": requesting timeout: " + timeMS + "ms for " + function + " in thread " + Thread.currentThread());
             TimerTask task = new TimerTask() {
                 @Override
